Fix ns_cvar_float returning int
From pierow

+++ main/source/mod/AvHServerUtil.cpp
@@ -1912,6 +1912,6 @@ float	ns_cvar_float(const cvar_t *cvar)
   {
     return 0;
   }
-  return atoi(cvar->string);
+  return atof(cvar->string);
 }
